I want the filenames of schemas to be lowercase.
I want the names of the schema’s to be lower case, so I renamed them.
However it seems that my file system (mac) and github are only
partially case sensitive. By deleting the files completely and then
re-inserting them I hope to convince the system that the files should
be lowercase.